backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pinecone-client==5.0.1
python-dotenv==1.0.0
pydantic[email]>=2.8.0
email-validator>=2.0.0
sentence-transformers>=2.7.0
transformers>=4.41.0
tokenizers>=0.20,<0.21
torch>=2.0.0
groq>=0.33.1
pymongo==4.6.1
motor==3.3.2
python-docx==1.1.0
striprtf==0.0.26
youtube-transcript-api==0.6.1
yt-dlp==2024.3.10
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
google-auth>=2.35.0
